rename mettagrid.config.config to mettagrid.base_config
