+++ common/src/metta/common/tool/run_tool.py
@@ -24,7 +24,7 @@ from metta.common.tool import Tool
 from metta.common.util.log_config import init_logging
 from metta.common.util.text_styles import bold, cyan, green, red, yellow
 from metta.rl.system_config import seed_everything
-from mettagrid.config import Config
+from mettagrid.base_config import Config
 from mettagrid.util.module import load_symbol
 
 logger = logging.getLogger(__name__)
