+++ metta/gridworks/configs/registry.py
@@ -10,7 +10,7 @@ from typing import Callable, Literal, cast, get_args
 
 from metta.common.util.fs import get_repo_root
 from metta.gridworks.configs.lsp import LSPClient
-from mettagrid.config import Config
+from mettagrid.base_config import Config
 from mettagrid.util.module import load_symbol
 
 logger = logging.getLogger(__name__)
