+++ metta/rl/loss/dynamics.py
@@ -10,7 +10,7 @@ from torch import Tensor
 from metta.agent.policy import Policy
 from metta.rl.loss import Loss
 from metta.rl.training import ComponentContext
-from mettagrid.config import Config
+from mettagrid.base_config import Config
 
 
 class DynamicsConfig(Config):
