+++ metta/rl/loss/loss_config.py
@@ -4,7 +4,7 @@ import torch
 from pydantic import Field
 
 from metta.agent.policy import Policy
-from mettagrid.config import Config
+from mettagrid.base_config import Config
 
 if TYPE_CHECKING:
     from metta.rl.training import TrainingEnvironment
