+++ metta/rl/loss/sl_kickstarter.py
@@ -11,7 +11,7 @@ from metta.agent.policy import Policy
 from metta.rl.loss import Loss
 from metta.rl.trainer_config import TrainerConfig
 from metta.rl.training import ComponentContext
-from mettagrid.config import Config
+from mettagrid.base_config import Config
 
 
 class SLKickstarterConfig(Config):
