+++ metta/rl/training/checkpointer.py
@@ -8,7 +8,7 @@ from pydantic import Field
 from metta.agent.policy import Policy, PolicyArchitecture
 from metta.rl.checkpoint_manager import CheckpointManager
 from metta.rl.training import DistributedHelper, EnvironmentMetaData, TrainerComponent
-from mettagrid.config import Config
+from mettagrid.base_config import Config
 
 logger = logging.getLogger(__name__)
