+++ metta/rl/training/evaluator.py
@@ -20,7 +20,7 @@ from metta.rl.evaluate import (
 from metta.rl.training import TrainerComponent
 from metta.sim.simulation_config import SimulationConfig
 from metta.tools.utils.auto_config import auto_replay_dir
-from mettagrid.config import Config
+from mettagrid.base_config import Config
 
 logger = logging.getLogger(__name__)
