+++ metta/rl/training/gradient_reporter.py
@@ -6,7 +6,7 @@ import torch
 from pydantic import Field
 
 from metta.rl.training import TrainerComponent
-from mettagrid.config import Config
+from mettagrid.base_config import Config
 
 logger = logging.getLogger(__name__)
