+++ metta/rl/training/stats_reporter.py
@@ -16,7 +16,7 @@ from metta.eval.eval_request_config import EvalRewardSummary
 from metta.rl.stats import accumulate_rollout_stats, compute_timing_stats, process_training_stats
 from metta.rl.training.component import TrainerComponent
 from metta.rl.utils import should_run
-from mettagrid.config import Config
+from mettagrid.base_config import Config
 
 logger = logging.getLogger(__name__)
