+++ metta/sweep/schedulers/batched_synced.py
@@ -15,7 +15,7 @@ from pydantic import Field
 
 from metta.adaptive.models import JobDefinition, JobStatus, RunInfo
 from metta.adaptive.utils import create_eval_job, create_training_job, generate_run_id
-from mettagrid.config import Config
+from mettagrid.base_config import Config
 
 logger = logging.getLogger(__name__)
