+++ packages/mettagrid/python/src/mettagrid/config/__init__.py
@@ -1,4 +1,3 @@
-from .config import Config
 from .mettagrid_c_config import from_mettagrid_config
 from .mettagrid_config import (
     ActionConfig,
@@ -24,7 +23,6 @@ __all__ = [
     "AssemblerConfig",
     "AttackActionConfig",
     "ChangeGlyphActionConfig",
-    "Config",
     "ConverterConfig",
     "from_mettagrid_config",
     "GameConfig",
