+++ packages/mettagrid/python/src/mettagrid/mapgen/scenes/random_scene.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from mettagrid.config.config import Config
+from mettagrid.base_config import Config
 from mettagrid.mapgen.scene import ChildrenAction, Scene, SceneConfig
