+++ packages/mettagrid/python/src/mettagrid/util/module.py
@@ -2,7 +2,7 @@ import importlib
 
 
 def load_symbol(full_name: str):
-    """Load a symbol from a full name, for example: 'mettagrid.config.Config' -> Config."""
+    """Load a symbol from a full name, for example: 'mettagrid.base_config.Config' -> Config."""
     parts = full_name.rsplit(".", 1)
     if len(parts) != 2:
         raise ValueError(f"Invalid symbol name: {full_name}")
